+++ VTK/Infovis/vtkContingencyStatistics.cxx
@@ -433,7 +433,7 @@ void vtkContingencyStatistics::ExecuteDerive( vtkDataObject* inMetaDO )
   
   // Resize output meta so marginal PDF tables can be appended
   unsigned int nBlocks = inMeta->GetNumberOfBlocks();
-  inMeta->SetNumberOfBlocks( nBlocks + marginalCounts.size() );
+  inMeta->SetNumberOfBlocks( nBlocks + static_cast<unsigned int>(marginalCounts.size()) );
 
   // Rows of the marginal PDF tables contain:
   // 0: variable value
